Make sure density shape is integral

+++ box_dft.py
@@ -134,7 +134,7 @@ if __name__ == "__main__":
     box_dims = np.array([8.0, 4.0, 1.0]) # angstroms
     points_per_angstrom = 20 # points per anstrom
     electron_count = 8
-    density = np.ones(points_per_angstrom*box_dims)
+    density = np.ones((points_per_angstrom*box_dims).astype(np.int_)) # future-proofing would do some checks here
     # fill in density with appropriate guess (uniform?)
     box_volume = np.prod(box_dims)
     density *= electron_count/box_volume
